260107_mysql/requirements.txt

Flask==3.0.0
mysql-connector-python==9.1.0
Werkzeug==3.0.1
